+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
-
 ## [0.1.0](https://github.com/zoltanlaczko/wp-zotya-htaccess-protect/tree/0.1.0) - 2020-05-20 ##
 * Initial release.
+
+## [0.2.0](https://github.com/zoltanlaczko/wp-zotya-htaccess-protect/tree/0.2.0) - 2020-06-02 ##
+* Add authorization fix option if you have loopback issues
+* refactor remove function